Change parsing of values
Plain values must not start with " (otherwise they must be quoted
strings) and with parenthesis (to not confuse with syntax of YARS-PG
and Cypher and to possibly allow for extension of PG syntax).

Single quoted strings are also removed because these are less likely
required and to be equal to JSON strings. Unicode escape sequences are
not included yet.